move array macros
