+++ addon/array/filter-by.js
@@ -1,14 +1,12 @@
-import { normalizeArray } from './utils';
+import { normalizeArray } from '../utils';
 
-/* jshint -W080 */
-const defaultValue = undefined;
-/* jshint +W080 */
+const defaultValue = [];
 
 export default function(...keys) {
   return normalizeArray(keys, { defaultValue }, (array, key, value) => {
     if (!key) {
       return defaultValue;
     }
-    return array.findBy(key, value);
+    return array.filterBy(key, value);
   });
 }
