+++ addon/array/find-by.js
@@ -1,12 +1,14 @@
-import { normalizeArray } from './utils';
+import { normalizeArray } from '../utils';
 
-const defaultValue = [];
+/* jshint -W080 */
+const defaultValue = undefined;
+/* jshint +W080 */
 
 export default function(...keys) {
   return normalizeArray(keys, { defaultValue }, (array, key, value) => {
     if (!key) {
       return defaultValue;
     }
-    return array.filterBy(key, value);
+    return array.findBy(key, value);
   });
 }
