+++ addon/array/index.js
@@ -2,6 +2,10 @@ export { default as any } from './any';
 export { default as compact } from './compact';
 export { default as concat } from './concat';
 export { default as every } from './every';
+export { default as filterBy } from './filter-by';
+export { default as filter } from './filter';
+export { default as findBy } from './find-by';
+export { default as find } from './find';
 export { default as includes } from './includes';
 export { default as indexOf } from './index-of';
 export { default as lastIndexOf } from './last-index-of';
